Fixed up Nav bar issue
moved MLH banner from the live-view.comp.html and placed it into app.component.html (allowing to adjust the banner and navbar). Changed width of navbar to 90% to avoid the overflow. When you zoom in, just when it looks like it's about to overflow, it switches to the movile version of the site. Also gave MLH banner z-index of 1 instead of 10,000 so it can stay behind the nav bar in mobile. Also, please don't use z-index of "10,000" lmao